Moved nft crawler into its own submodule
